+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Changed
 
 - Add version to `auditable_index`
   [#427](https://github.com/collectiveidea/audited/pull/427)
+- Rename audited resource revision `version` attribute to `audit_version` and deprecate `version` attribute
+  [#443](https://github.com/collectiveidea/audited/pull/443)
 
 Fixed
 
@@ -38,7 +40,8 @@ Added
 
 Changed
 
-- None
+- Change/deprecate `version` attribute to `audit_version` attribute
+  [#443](https://github.com/collectiveidea/audited/pull/443)
 
 Fixed